Moved utility functions into a dedicated `helpers.rs` file
