+++ engine/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod board;
 mod moves;
+mod helpers;
